requirements.txt
flask 
pymongo
flask-jwt-extended
passlib[bcrypt]
python-dotenv